perf: cache GPU stats and pace the system-stats sampler to subscribers
